requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
asttokens==2.4.1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Deprecated==1.2.15
devtools==0.12.2
distro==1.9.0
eval_type_backport==0.2.0
executing==2.1.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
googleapis-common-protos==1.66.0
griffe==1.5.1
groq==0.13.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
jiter==0.8.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
logfire==2.6.2
logfire-api==2.6.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.15.2
numpy==2.0.0
openai==1.57.0
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==5.29.1
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic-ai==0.0.10
pydantic-ai-slim==0.0.10
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.40.2
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
wrapt==1.17.0
zipp==3.21.0
youtube-transcript-api==0.6.3
clipboard==0.0.4
pyperclip==1.9.0
st-copy-to-clipboard==0.1.6